decouple tab content from tab component itself
